Fixed bug on RID serialization: cluster position was casted to short! Adjusted tree leaf to use INT instead of SHORT for page sizes.

+++ core/src/main/java/com/orientechnologies/orient/core/id/ORecordId.java
@@ -171,7 +171,7 @@ public class ORecordId implements ORID, Comparable<ORecordId>, Comparator<ORecor
 	public byte[] toStream() {
 		byte[] buffer = new byte[PERSISTENT_SIZE];
 		OBinaryProtocol.short2bytes((short) clusterId, buffer, 0);
-		OBinaryProtocol.long2bytes((short) clusterPosition, buffer, OConstants.SIZE_SHORT);
+		OBinaryProtocol.long2bytes(clusterPosition, buffer, OConstants.SIZE_SHORT);
 		return buffer;
 	}
